pyproject.toml
[tool.poetry]
name = "lgpdnowgpt"
version = "0.1.0"
description = "GptPDF"
authors = ["maellson <mqmaellson39@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.13"
streamlit = "^1.31.1"
python-dotenv = "^1.0.1"
langchain = "^0.1.7"
langchain-openai = "^0.0.6"
pypdf2 = "^3.0.1"
faiss-cpu = "^1.7.4"
streamlit-chat = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
